fix(ci): python-full-coverage - split oversized executor test shard
